Remove from cart functionality

+++ ecommerce/views.py
@@ -91,8 +91,8 @@ def updateItem(request):
         data=json.loads(request.body)
         productId = data['productId']
         action = data['action']
-        #print('Action:', action)
-        #print('Product:', productId)
+        print('Action:', action)
+        print('Product:', productId)
 
         u = request.user
         customer = Customer.objects.get(user_id=u.id)
@@ -151,11 +151,22 @@ def processOrder(request):
         return redirect('login')
 
 @login_required
-def deletefromcart(request,id):
-    Order.objects.filter(id=id).delete()
-    messages.success(request, "Your item deleted form Shopcart.")
-    return render(request, 'ecommerce/checkout.html', context)
+def deletefromcart(request):
+    order, created = Order.objects.get_or_create(customer=customer, complete=False)
+    items = order.orderitem_set.all()
+    cartItems = order.get_cart_items
+    context = {'items':items, 'order':order, 'cartItems':cartItems}
+    
+    if request.is_ajax and request.method == "GET":
+        # get the nick name from the client side.
+        product_id = request.GET.get("product_id", None)
+        # check for the nick name in the database.
+        Order.objects.filter(id=id).delete()
+        #messages.success(request, "Your item deleted form Shopcart.")
+        return JsonResponse({'message': 'Item removed.'}, safe=False)
+    return JsonResponse({}, status = 400)
 
+        
 def signup(request):
 
     if request.method == 'POST':
